moving code to list_by_date to entity
