+++ backend/corpora/common/entities/project.py
@@ -1,8 +1,4 @@
-import typing
 import uuid
-from datetime import datetime
-
-from sqlalchemy import and_
 
 from .entity import Entity
 from ..corpora_orm import DbProject, DbProjectLink, ProjectStatus
@@ -10,6 +6,7 @@ from ..corpora_orm import DbProject, DbProjectLink, ProjectStatus
 
 class Project(Entity):
     table = DbProject
+    list_attributes = (DbProject.id, DbProject.created_at)
 
     def __init__(self, db_object: DbProject):
         super().__init__(db_object)
@@ -69,40 +66,6 @@ class Project(Entity):
 
     @classmethod
     def list_projects_in_time_range(cls, *args, **kwargs):
-        return cls.list_in_time_range(*args, filters=[DbProject.status == ProjectStatus.LIVE.name], **kwargs)
+        return cls.list_attributes_in_time_range(*args, filters=[DbProject.status == ProjectStatus.LIVE.name], **kwargs)
 
     @classmethod
-    def list_in_time_range(
-        cls, to_date: int = None, from_date: int = None, filters: list = None
-    ) -> typing.List[typing.Dict]:
-        """
-        Queries the database for projects that have been created within the specified time range.
-
-        :param to_date: If provided, only lists projects that were created before this date. Format of param is Unix
-        timestamp since the epoch in UTC timezone.
-        :param from_date: If provided, only lists projects that were created after this date. Format of param is Unix
-        timestamp since the epoch in UTC timezone.
-        :param filters: additional filters to apply to the query.
-        :return: The results is a list of flattened dictionaries containing the `list_entities`
-        """
-
-        filters = filters if filters else []
-        list_entities = [DbProject.created_at, DbProject.id]
-
-        def to_dict(db_object):
-            _result = {}
-            for _field in db_object._fields:
-                _result[_field] = getattr(db_object, _field)
-            return _result
-
-        if to_date:
-            filters.append(DbProject.created_at <= datetime.fromtimestamp(to_date))
-        if from_date:
-            filters.append(DbProject.created_at >= datetime.fromtimestamp(from_date))
-
-        results = [
-            to_dict(result)
-            for result in cls.db.session.query(DbProject).with_entities(*list_entities).filter(and_(*filters)).all()
-        ]
-
-        return results
